requirements.txt
groq>=0.4.0
pandas>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
openpyxl>=3.1.0
